--- src/components/AccountHeader/style.tsx
@@ -1,20 +1,0 @@
-import styled from "styled-components";
-
-export const EnvironmentGroup = styled.div`
-    font-weight: bold;
-    width: 10%;
-`
-
-export const PlatformGroup = styled.div`
-    font-weight: bold;
-    width: 15%;
-`
-
-export const UserIdGroup = styled.div`
-    font-weight: bold;
-`
-
-export const ButtonGroup = styled.div`
-    float: right;
-    gap: 10px;
-`
