--- src/components/AccountRow/style.tsx
@@ -1,16 +1,0 @@
-import styled from "styled-components";
-
-export const EnvironmentGroup = styled.div`
-    width: 10%;
-    color: #494949;
-`
-
-export const PlatformGroup = styled.div`
-    width: 15%;
-    color: #494949;
-`
-
-export const ButtonGroup = styled.div`
-    float: right;
-    gap: 10px;
-`
